*** empty log message ***
git-svn-id: http://talendforge.org/svn/tos/trunk@219 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.designer.runprocess/src/main/java/org/talend/designer/runprocess/ui/views/ProcessView.java
@@ -29,7 +29,6 @@ import org.eclipse.jface.action.IMenuManager;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.FocusEvent;
 import org.eclipse.swt.events.FocusListener;
-import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.FontData;
 import org.eclipse.swt.layout.GridData;
@@ -228,7 +227,9 @@ public class ProcessView extends ViewPart {
 
         @Override
         public void run() {
-            processComposite2.exec();
+            if (processComposite2.hasProcess()) {
+                processComposite2.exec();
+            }
         }
 
     }
@@ -251,7 +252,9 @@ public class ProcessView extends ViewPart {
 
         @Override
         public void run() {
-            processComposite2.debug();
+            if (processComposite2.hasProcess()) {
+                processComposite2.debug();
+            }
         }
 
     }
@@ -274,7 +277,9 @@ public class ProcessView extends ViewPart {
 
         @Override
         public void run() {
-            processComposite2.kill();
+            if (processComposite2.hasProcess()) {
+                processComposite2.kill();
+            }
         }
 
     }
